Refactor `WhereFilterIntersection` into `semantic_layer_common.py`
This is useful because moving it out of `nodes.py` avoids circular imports.
