+++ tests/unit/test_manifest.py
@@ -26,11 +26,10 @@ from dbt.contracts.graph.nodes import (
     Metric,
     MetricInputMeasure,
     MetricTypeParams,
-    WhereFilter,
-    WhereFilterIntersection,
     Group,
     RefArgs,
 )
+from dbt.contracts.graph.semantic_layer_common import WhereFilter, WhereFilterIntersection
 from dbt.contracts.graph.unparsed import (
     ExposureType,
     Owner,
